Перенёс Composable функции в отдельные файлы из фрагментов
